third_party/WebKit/LayoutTests/external/wpt/css/css-tables/html5-table-formatting-3-expected.txt

This is a testharness.js-based test.
PASS Control test for table-cell padding and border-spacing, etc (width)
PASS Control test for table-cell padding and border-spacing (height)
PASS Anonymous consecutive columns spanned by the same set of cells are merged
FAIL Anonymous consecutive rows spanned by the same set of cells are merged assert_equals: expected 100 but got 75
FAIL Explicitely-defined consecutive columns spanned by the same set of cells are not merged assert_equals: expected 75 but got 50
PASS Explicitely-defined consecutive rows spanned by the same set of cells are not merged
FAIL Explicitely-defined consecutive columns spanned by the same set of cells are not merged, and cells span across border-spacing assert_equals: expected 12.5 but got 0
FAIL Explicitely-defined consecutive rows spanned by the same set of cells are not merged, and cells span across border-spacing assert_equals: expected 12.5 but got 12
Harness: the test ran to completion.
